modified:   src/components/Common/Table/GenericTable.tsx 	new file:   src/hooks/useDataTable.tsx 	modified:   src/pages/AdminManageCategoryPage.tsx
